+++ sam/metrics/__init__.py
@@ -4,6 +4,7 @@ from .mase import mean_absolute_scaled_error
 from .keras_metrics import keras_tilted_loss, keras_joint_mse_tilted_loss, \
     keras_rmse, get_keras_forecasting_metrics
 from .train_mean_r2 import train_mean_r2
+from .custom_callbacks import R2Evaluation
 
 from . import mase
 from . import keras_metrics