+++ sam/visualization/tests/test_flatline_diagnostic_plot.py
@@ -8,7 +8,7 @@ from sam.visualization import diagnostic_flatline_removal
 
 class TestFlatlineDiagnosticPlot(unittest.TestCase):
 
-    @pytest.mark.mpl_image_compare(tolerance=20)
+    @pytest.mark.mpl_image_compare(tolerance=30)
     def test_flatline_removal(self):
 
         # create some random data