+++ sam/visualization/tests/test_incident_heatmap.py
@@ -8,7 +8,7 @@ import seaborn as sns
 
 class TestIncidentHeatmap(unittest.TestCase):
 
-    @pytest.mark.mpl_image_compare(tolerance=20)
+    @pytest.mark.mpl_image_compare(tolerance=30)
     def test_incident_heatmap(self):
         range = pd.date_range('1/1/2011', periods=45, freq='D')
         ts = pd.DataFrame({'incident': np.tile([0.1, 0.2, 0.3], 15),
@@ -20,7 +20,7 @@ class TestIncidentHeatmap(unittest.TestCase):
                                    figsize=(15, 8), datefmt="%Y-%m-%d")
         return ax.get_figure()
 
-    @pytest.mark.mpl_image_compare(tolerance=20)
+    @pytest.mark.mpl_image_compare(tolerance=30)
     def test_incident_heatmap_advanced(self):
         range = pd.date_range('1/1/2011', periods=45, freq='D')
         ts = pd.DataFrame({'incident': np.tile([0.1, 0.2, 0.3], 15),