+++ sam/visualization/tests/test_threshold_plot.py
@@ -6,14 +6,14 @@ import numpy as np
 
 class TestThresholdlot(unittest.TestCase):
 
-    @pytest.mark.mpl_image_compare(tolerance=20)
+    @pytest.mark.mpl_image_compare(tolerance=30)
     def test_regular_threshold_plot(self):
         y_true = np.array([0, 1, 0, 1, 0, 1])
         y_scores = np.array([0.1, 0.2, 0.3, 0.4, 0.5, 0.6])
         ax = plot_threshold_curve(y_true, y_scores)
         return ax.get_figure()
 
-    @pytest.mark.mpl_image_compare(tolerance=20)
+    @pytest.mark.mpl_image_compare(tolerance=30)
     def test_incident_threshold_plot(self):
         y_scores = np.array([0.1, 0.2, 0.3, 0.4, 0.5, 0.6])
         y_incidents = np.array([0, 0, 1, 0, 0, 1])